Changed data types for Latitude and Longtitude in Place model from float to decimal for better accuracy

+++ uTraverse.PlacesAPI/Models/Place.cs
@@ -10,7 +10,7 @@ public class Place
 
     public required string Address { get; set; }
 
-    public float Latitude { get; set; }
+    public decimal Latitude { get; set; }
 
-    public float Longitude { get; set; }
+    public decimal Longitude { get; set; }
 }
